+++ packages/forma-36-react-components/src/components/Dropdown/Dropdown/Dropdown.tsx
@@ -15,7 +15,7 @@ export type positionType =
   | 'top-left';
 
 export type DropdownProps = {
-  toggleElement?: React.ReactNode;
+  toggleElement?: React.ReactElement;
   submenuToggleLabel?: string;
   position: positionType;
   isOpen: boolean;
@@ -26,6 +26,7 @@ export type DropdownProps = {
   className?: string;
   children: React.ReactNode;
   isFullWidth?: boolean;
+  isAutoalignmentEnabled?: boolean;
 } & typeof defaultProps;
 
 export interface AnchorDimensionsAndPositonType {
@@ -46,6 +47,7 @@ const defaultProps = {
   testId: 'cf-ui-dropdown',
   position: 'bottom-left',
   isOpen: false,
+  isAutoalignmentEnabled: true,
   getContainerRef: () => {},
 };
 
@@ -171,6 +173,7 @@ export class Dropdown extends Component<DropdownProps, DropdownState> {
       dropdownContainerClassName,
       children,
       isOpen,
+      isAutoalignmentEnabled,
       ...otherProps
     } = this.props;
 
@@ -184,7 +187,11 @@ export class Dropdown extends Component<DropdownProps, DropdownState> {
         onLeave={() => this.openMenu(false)}
         {...otherProps}
       >
-        {toggleElement}
+        {toggleElement &&
+          React.cloneElement(toggleElement, {
+            'aria-haspopup': 'listbox',
+            'aria-expanded': this.state.isOpen,
+          })}
         {this.state.isOpen && (
           <DropdownContainer
             anchorDimensionsAndPositon={this.state.anchorDimensionsAndPositon}
@@ -212,7 +219,11 @@ export class Dropdown extends Component<DropdownProps, DropdownState> {
         }}
         {...otherProps}
       >
-        {toggleElement}
+        {toggleElement &&
+          React.cloneElement(toggleElement, {
+            'aria-haspopup': 'listbox',
+            'aria-expanded': this.state.isOpen,
+          })}
         {this.state.isOpen && (
           <DropdownContainer
             className={dropdownContainerClassName}
@@ -220,6 +231,7 @@ export class Dropdown extends Component<DropdownProps, DropdownState> {
             submenu={false}
             width={this.state.containerWidth}
             dropdownAnchor={this.dropdownAnchor}
+            isAutoalignmentEnabled={isAutoalignmentEnabled}
             anchorDimensionsAndPositon={this.state.anchorDimensionsAndPositon}
             onClose={this.props.onClose}
             openSubmenu={this.openSubmenu}