+++ packages/forma-36-react-components/src/components/Dropdown/DropdownList/DropdownList.tsx
@@ -37,6 +37,7 @@ export class DropdownList extends Component<DropdownListProps> {
       <ul
         ref={listRef}
         data-test-id={testId}
+        role="listbox"
         style={{
           maxHeight: maxHeight || 'auto',
           overflowY: maxHeight ? 'auto' : 'visible',