+++ packages/components/forma-36-react-datepicker/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.1-alpha.7](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.1.1-alpha.6...@contentful/forma-36-react-datepicker@0.1.1-alpha.7) (2019-11-21)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+## [0.1.1-alpha.6](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.1.1-alpha.5...@contentful/forma-36-react-datepicker@0.1.1-alpha.6) (2019-11-20)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
 ## [0.1.1-alpha.5](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.1.1-alpha.4...@contentful/forma-36-react-datepicker@0.1.1-alpha.5) (2019-11-18)
 
 **Note:** Version bump only for package @contentful/forma-36-react-datepicker